backend/requirements.txt
django==4.1.4
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
drf-yasg==1.21.4
django-rest-knox==4.2.0